Dependencies: Add support for Python 3.12
Python 3.12 support was added to `aiida-core` with v2.5.0. This version
upgraded to SQLAlchemy v2 which as of v2.0.19 started emitting a warning.
This is ignored by a filter in `aiida-core`, but this is made undone by
`pytest` so it has to be ignored again explicitly in the `pytest` config
in the `pyproject.toml`.