`PseudoPotentialData`: Base hash only on file content and element (#168)
By default, the hash for data nodes is computed from a set of attributes
that includes the version of `aiida-core` as well as the plugin package
that provides the data plugin, if applicable. This means that the hash
of a data node changes merely if the version of `aiida-core` or the
plugin package changes, even if its content is identical.

This choice was made in `aiida-core` to err on the safe side. However,
for pseudopotentials, this is unnecessary and actually reduces the
effectiveness of caching. If a calculation is run with pseudos that has
already run before, but the version of `aiida-pseudo` is different with
which the pseudo was installed, the calculation will not be cached but
run again.

The `PseudoPotentialData` now customizes the caching class, ensuring the
hash is calculated just based on the element and the file content. This
should uniquely define a pseudopotential and so should be safe enough to
base the hash upon.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>